# src/presentation/http/digest_subscription_handler.rs
//! DigestSubscription REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::{ApiResponse, BackboneCrudHandler};

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{DigestSubscriptionService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateDigestSubscriptionDto, UpdateDigestSubscriptionDto, PatchDigestSubscriptionDto, DigestSubscriptionResponseDto};

use crate::domain::state_machine::{digest_subscription_stateState, digest_subscription_stateStateMachine, digest_subscription_stateTransition};

/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum DigestSubscriptionError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for DigestSubscriptionError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for DigestSubscriptionError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "DIGESTSUBSCRIPTION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "DIGESTSUBSCRIPTION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "DIGESTSUBSCRIPTION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "DIGESTSUBSCRIPTION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for DigestSubscription.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /digest_subscriptions | List with pagination |
/// | POST | /digest_subscriptions | Create new |
/// | GET | /digest_subscriptions/:id | Get by ID |
/// | PUT | /digest_subscriptions/:id | Full update |
/// | PATCH | /digest_subscriptions/:id | Partial update |
/// | DELETE | /digest_subscriptions/:id | Soft delete |
/// | POST | /digest_subscriptions/bulk | Bulk create |
/// | POST | /digest_subscriptions/upsert | Upsert |
/// | GET | /digest_subscriptions/trash | List deleted |
/// | POST | /digest_subscriptions/:id/restore | Restore |
/// | DELETE | /digest_subscriptions/empty | Empty trash |
/// | GET | /digest_subscriptions/:id/deleted | Get deleted by ID |
/// | DELETE | /digest_subscriptions/trash/:id | Permanent delete from trash |
/// | GET | /digest_subscriptions/count | Count active entities |
/// | GET | /digest_subscriptions/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(DigestSubscriptionService::with_repository(repository));
/// let router = create_digest_subscription_routes(service);
/// ```
pub fn create_digest_subscription_routes(service: Arc<DigestSubscriptionService>) -> Router {
    BackboneCrudHandler::<DigestSubscriptionService, DigestSubscription, CreateDigestSubscriptionDto, UpdateDigestSubscriptionDto, DigestSubscriptionResponseDto>::routes(
        service,
        "/digest_subscriptions",
    )
}

/// Create Axum router with only the read (GET) endpoints for DigestSubscription.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_digest_subscription_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_digest_subscription_read_routes(service: Arc<DigestSubscriptionService>) -> Router {
    BackboneCrudHandler::<DigestSubscriptionService, DigestSubscription, CreateDigestSubscriptionDto, UpdateDigestSubscriptionDto, DigestSubscriptionResponseDto>::read_routes(
        service,
        "/digest_subscriptions",
    )
}

/// Create Axum router with only the write (mutation) endpoints for DigestSubscription.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_digest_subscription_write_routes(service: Arc<DigestSubscriptionService>) -> Router {
    BackboneCrudHandler::<DigestSubscriptionService, DigestSubscription, CreateDigestSubscriptionDto, UpdateDigestSubscriptionDto, DigestSubscriptionResponseDto>::write_routes(
        service,
        "/digest_subscriptions",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_digest_subscription_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<DigestSubscriptionService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_digest_subscription_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}

// =============================================================================
// State Transition Handlers
// =============================================================================

/// Execute unsubscribe transition on a DigestSubscription.
///
/// POST /digest_subscriptions/:id/transitions/unsubscribe
pub async fn unsubscribe_transition(
    axum::extract::State(service): axum::extract::State<Arc<DigestSubscriptionService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<DigestSubscriptionResponseDto>::not_found("DigestSubscription", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<DigestSubscriptionResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = digest_subscription_stateTransition::Unsubscribe.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "digest_subscription:transition:unsubscribe");
        let has_update_perm = auth.permissions.iter().any(|p| p == "digest_subscription:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<DigestSubscriptionResponseDto>::error("Insufficient permissions for unsubscribe transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: digest_subscription_stateState = entity.state.to_string().parse()
        .unwrap_or(digest_subscription_stateState::default());
    let sm = digest_subscription_stateStateMachine::from_state(current_state);
    if !sm.can_transition(digest_subscription_stateTransition::Unsubscribe) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<DigestSubscriptionResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Unsubscribed".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: DigestSubscriptionResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<DigestSubscriptionResponseDto>::not_found("DigestSubscription", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<DigestSubscriptionResponseDto>::error(e.to_string()))),
    }
}

/// Execute resubscribe transition on a DigestSubscription.
///
/// POST /digest_subscriptions/:id/transitions/resubscribe
pub async fn resubscribe_transition(
    axum::extract::State(service): axum::extract::State<Arc<DigestSubscriptionService>>,
    axum::extract::Path(id): axum::extract::Path<String>,
    #[cfg(feature = "auth")] axum::Extension(auth): axum::Extension<AuthContext>,
) -> impl axum::response::IntoResponse {
    use axum::{http::StatusCode, Json};

    // Get current entity
    let entity = match service.get_by_id(&id).await {
        Ok(Some(e)) => e,
        Ok(None) => return (StatusCode::NOT_FOUND, Json(ApiResponse::<DigestSubscriptionResponseDto>::not_found("DigestSubscription", &id))),
        Err(e) => return (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<DigestSubscriptionResponseDto>::error(e.to_string()))),
    };

    // Check permission (if auth enabled)
    #[cfg(feature = "auth")]
    {
        let allowed_roles = digest_subscription_stateTransition::Resubscribe.allowed_roles();
        let has_specific_perm = auth.permissions.iter().any(|p| p == "digest_subscription:transition:resubscribe");
        let has_update_perm = auth.permissions.iter().any(|p| p == "digest_subscription:update");
        if !has_specific_perm && !has_update_perm {
            return (StatusCode::FORBIDDEN, Json(ApiResponse::<DigestSubscriptionResponseDto>::error("Insufficient permissions for resubscribe transition")));
        }
    }

    // Create state machine from entity's actual status and validate transition
    let current_state: digest_subscription_stateState = entity.state.to_string().parse()
        .unwrap_or(digest_subscription_stateState::default());
    let sm = digest_subscription_stateStateMachine::from_state(current_state);
    if !sm.can_transition(digest_subscription_stateTransition::Resubscribe) {
        return (StatusCode::BAD_REQUEST, Json(ApiResponse::<DigestSubscriptionResponseDto>::error("Transition not allowed from current state")));
    }

    // Apply transition via partial update
    let mut fields: HashMap<String, serde_json::Value> = HashMap::new();
    fields.insert("state".to_string(), serde_json::Value::String("Subscribed".to_string()));

    match service.partial_update(&id, fields).await {
        Ok(Some(updated)) => {
            let response: DigestSubscriptionResponseDto = updated.into();
            (StatusCode::OK, Json(ApiResponse::ok(response)))
        }
        Ok(None) => (StatusCode::NOT_FOUND, Json(ApiResponse::<DigestSubscriptionResponseDto>::not_found("DigestSubscription", &id))),
        Err(e) => (StatusCode::INTERNAL_SERVER_ERROR, Json(ApiResponse::<DigestSubscriptionResponseDto>::error(e.to_string()))),
    }
}

/// Create routes for state transitions.
pub fn create_digest_subscription_transition_routes(service: Arc<DigestSubscriptionService>) -> Router {
    use axum::routing::post;

    Router::new()
        .route("/digest_subscriptions/:id/transitions/unsubscribe", post(unsubscribe_transition))
        .route("/digest_subscriptions/:id/transitions/resubscribe", post(resubscribe_transition))
        .with_state(service)
}